backend/pyproject.toml
[tool.poetry]
name = "lunch"
version = "0.1.0"
description = ""
authors = ["nahoon kang <ejrtks102020@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
fastapi = "^0.104.1"
sqlalchemy = "^2.0.28"
pydantic-settings = "^2.2.1"
pydantic = {extras = ["email"], version = "^2.7.0"}
alembic = "^1.13.1"
loguru = "^0.7.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
jwt = "^1.3.1"
uvicorn = "^0.29.0"
asyncmy = "^0.2.9"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
greenlet = "^3.0.3"
bcrypt = "^4.1.2"
icecream = "^2.1.3"
sqladmin = "^0.19.0"
aiosqlite = "^0.20.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
